updated exception handling according to sir's latest guidelines

+++ feedback/src/main/java/com/feedback/service/impl/FeedbackServiceImpl.java
@@ -17,34 +17,76 @@ public class FeedbackServiceImpl implements FeedbackService {
 	@Autowired
 	private FeedbackDAO dao;
 
+	private boolean isValidStarsOverall(int starsOverall) {
+		if (starsOverall < 0 || starsOverall > 10) {
+			return false;
+		}
+		return true;
+	}
+
+	private boolean isValidstarsAuthor(int starsAuthor) {
+		if (starsAuthor < 0 || starsAuthor > 10) {
+			return false;
+		}
+		return true;
+	}
+
 	@Override
-	public Feedback createFeedback(Feedback Feedback) {
+	public Feedback createFeedback(Feedback Feedback) throws BusinessException {
+		if (!isValidStarsOverall(Feedback.getStarsOverall())) {
+			throw new BusinessException("starsOverall should be between 0 and 10");
+		}
+		if (!isValidstarsAuthor(Feedback.getStarsAuthor())) {
+			throw new BusinessException("starsAuthor should be between 0 and 10");
+		}
 		return dao.save(Feedback);
 
 	}
 
 	@Override
 	public Feedback getFeedbackById(int id) throws BusinessException {
+		if (id < 0) {
+			throw new BusinessException("id cannot be" + id);
+		}
 		Feedback feedback = null;
 		try {
-		feedback = dao.findById(id).get();
-		}catch (NoSuchElementException e) {
+			feedback = dao.findById(id).get();
+		} catch (NoSuchElementException e) {
 			throw new BusinessException("No feedback found for id " + id);
 		}
 		return feedback;
 	}
 
 	@Override
-	public Feedback updateFeedback(Feedback Feedback) {
+	public Feedback updateFeedback(Feedback Feedback) throws BusinessException {
+
+		if (!isValidStarsOverall(Feedback.getStarsOverall())) {
+			throw new BusinessException("starsOverall should be between 0 and 10");
+		}
+		if (!isValidstarsAuthor(Feedback.getStarsAuthor())) {
+			throw new BusinessException("starsAuthor should be between 0 and 10");
+		}
+
+		if (Feedback.getId() < 0) {
+			throw new BusinessException("id cannot be " + Feedback.getId());
+		}
+		try {
+			dao.findById(Feedback.getId()).get();
+		} catch (NoSuchElementException e) {
+			throw new BusinessException("No feedback found for id " + Feedback.getId());
+		}
 		return dao.save(Feedback);
 
 	}
 
 	@Override
 	public void deleteFeedback(int id) throws BusinessException {
+		if (id < 0) {
+			throw new BusinessException("id cannot be" + id);
+		}
 		try {
 			dao.deleteById(id);
-		}catch (EmptyResultDataAccessException e) {
+		} catch (EmptyResultDataAccessException e) {
 			throw new BusinessException("No feedback found for id " + id);
 		}
